backend/requirements/base.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-redis>=5.4,<6.0
django-filter>=24.0
dj-database-url>=2.1,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# Task Queue
django-q2>=1.6,<2.0

# LangChain & AI
langchain>=0.3,<1.0
langchain-google-genai>=2.0,<3.0
langchain-community>=0.3,<1.0
langchain-chroma>=0.2,<1.0
langchain-groq>=0.1,<1.0
langchain-openai>=0.1,<1.0
chromadb>=0.5,<1.0
groq>=0.9,<1.0

# Document Processing
PyPDF2>=3.0,<4.0
PyMuPDF>=1.24,<2.0
python-docx>=1.1,<2.0
openpyxl>=3.1,<4.0
pandas>=2.2,<3.0
python-magic>=0.4,<1.0

# Utilities
python-decouple>=3.8,<4.0
structlog>=24.0
tiktoken>=0.7
hashlib-additional>=1.0; python_version < "3.0"

# API Documentation
drf-spectacular>=0.27,<1.0

# Logging & Monitoring
sentry-sdk[django]>=2.0,<3.0
